requirements.txt
llama-index
pydantic
pyyaml
openai
chainlit
pydantic-ai 
wikipedia  
